Revert "Add basic JIT"
This reverts commit d139698e089fc219f9f135d5cd43de9e327dd5a4.

+++ src/frontend/ast.rs
@@ -92,8 +92,8 @@ impl std::cmp::PartialEq for InterpEntry {
 
 #[derive(PartialEq, Debug, Clone)]
 pub enum ExprKind {
-    Float(f64),
-    Int(i32),
+    Float(f32),
+    Int(i64),
     String(String),
     InterpStr(String, Vec<InterpEntry>),
     Bool(bool),
